# TypeScript: faster getExportsForFile