TL;DR

  * Don't use createProgram because
    * we've already read the file
    * we don't need dependents
  * Don't use TypeChecker because
    * undocumented `node.name.text` gives us exactly what `getSymbolAtLocation(node.name).getName()` gave us (as far as I can tell!)

Seen 966ms => 21ms and 661ms => 2ms reductions on a small project, see https://github.com/electron/electron-compile/issues/224#issuecomment-297523068